// frontend/src/api/generated/models/factoryDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export interface FactoryDto {
  id: number;
  name: string;
  timeZone: string;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}