// frontend/src/api/generated/factories/factories.zod.ts
export const PostApiFactoriesWithJsonBody = zod.object({
  "name": zod.string(),
  "timeZone": zod.string()
})
// ...
export const PutApiFactoriesIdWithJsonBody = zod.object({
  "name": zod.string(),
  "timeZone": zod.string()
})